# backend/api/docs/swagger.yaml
basePath: /
definitions:
...
  echo-backend_internal_models_llmops.PromptTemplate:
    properties:
      active_version:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  echo-backend_internal_models_llmops.PromptVersion:
    properties:
      bound_tools:
        items:
          type: string
        type: array
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      status:
        type: string
      system_prompt:
        type: string
      template_id:
        type: string
      variables:
        items:
          type: string
        type: array
      version:
        type: integer
    type: object
...
  echo-backend_internal_service_llmops.PlaygroundRequest:
    properties:
      features:
        items:
          type: string
        type: array
      models:
        items:
          type: string
        type: array
      prompt:
        type: string
      skills:
        items:
          type: string
        type: array
      variables:
        additionalProperties:
          type: string
        type: object
    type: object
...
  internal_handler_llmops.createTemplateReq:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  internal_handler_llmops.createVersionReq:
    properties:
      bound_tools:
        items:
          type: string
        type: array
      system_prompt:
        type: string
      variables:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@echo.app
    name: Echo API Support
    url: https://echo.app/support
  description: Core API Server for Echo Platform providing Auth, Chat, Sessions, Models,
    Settings, Admin, Memory, and System Services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://echo.app/terms
  title: Echo Backend API
  version: "1.0"
paths:
...
  /api/v1/studio/playground:
    post:
      consumes:
      - application/json
      description: Streams multi-model prompt comparison results as Server-Sent Events
      parameters:
      - description: Playground payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/echo-backend_internal_service_llmops.PlaygroundRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: Event stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Run LLM playground
      tags:
      - Studio
  /api/v1/studio/prompts:
    get:
      description: Returns all prompt templates (LLMOps Studio)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List prompt templates
      tags:
      - Studio
    post:
      consumes:
      - application/json
      description: Creates a new prompt template (LLMOps Studio)
      parameters:
      - description: Template payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_llmops.createTemplateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/echo-backend_internal_models_llmops.PromptTemplate'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a prompt template
      tags:
      - Studio
  /api/v1/studio/prompts/{id}/promote/{version}:
    post:
      description: Promotes a version to production (LLMOps Studio)
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Promote a prompt template version
      tags:
      - Studio
  /api/v1/studio/prompts/{id}/rollback/{version}:
    post:
      description: Rolls back the production version of a prompt template (LLMOps
        Studio)
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Version number
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rollback a prompt template version
      tags:
      - Studio
  /api/v1/studio/prompts/{id}/versions:
    get:
      description: Returns the version history of a prompt template (LLMOps Studio)
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List prompt template versions
      tags:
      - Studio
    post:
      consumes:
      - application/json
      description: Creates a new version of a prompt template (LLMOps Studio)
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Version payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_llmops.createVersionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/echo-backend_internal_models_llmops.PromptVersion'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a prompt template version
      tags:
      - Studio
  /api/v1/studio/prompts/{id}/versions/{v}:
    get:
      description: Returns a specific version of a prompt template (LLMOps Studio)
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Version number
        in: path
        name: v
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo-backend_internal_models_llmops.PromptVersion'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a prompt template version
      tags:
      - Studio
  /api/v1/studio/prompts/active:
    get:
      description: Returns the active production version of a prompt template (LLMOps
        Studio)
      parameters:
      - description: Template name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo-backend_internal_models_llmops.PromptVersion'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get active prompt
      tags:
      - Studio
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authorization header (Format: Bearer <JWT>).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"